Allow Batch job modules to use other Batch job modules
Used jobs are copied into the dependencies of the using job, signaling
Batch that the using job cannot be scheduled until the used job(s)
succeed. As a result the toolkit can submit multiple jobs, even if
they are not independent, and as long as the "used" fields are
configured correctly, Batch will schedule and run them in the
correct order.